space/requirements.txt
fastapi
uvicorn
torch
transformers==4.44.2
Pillow
python-multipart
einops
timm
